// src/core/result/types/http-status-codes.enum.ts
export enum HTTP_STATUS_CODES {
  OK_200 = 200,
  CREATED_201 = 201,
  NO_CONTENT_204 = 204,

  BAD_REQUEST_400 = 400,
  UNAUTHORIZED_401 = 401,
  FORBIDDEN_403 = 403,
  NOT_FOUND_404 = 404,
  NOT_ALLOWED_405 = 405,
  CONFLICT_409 = 409,
  UNPROCESSABLE_ENTITY_422 = 422,

  INTERNAL_SERVER_ERROR_500 = 500,
}